requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langgraph-checkpoint-sqlite>=1.0.0

# LLM Provider - OpenAI
openai>=1.0.0
langchain-openai>=0.0.2

# Telegram Bot (DEPRECATED — will be removed in a future release)
python-telegram-bot>=21.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration
python-dotenv>=1.0.0

# Storage & Database
sqlalchemy>=2.0.0
aiosqlite>=0.21.0

# HTTP & Web Scraping
httpx>=0.25.0
beautifulsoup4>=4.12.0

# Google APIs (Gmail)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

pydantic>=2.7.0

qdrant-client>=1.3.0

# Calendar
caldav>=1.3.0
icalendar>=5.0.0
pytz
dateparser
